grid_bot_v2/requirements.txt
pybit
pandas
python-dotenv
requests
sqlalchemy
lightgbm
scikit-learn
numpy
scipy
